+++ src/inet/linklayer/configurator/gatescheduling/common/EagerGateScheduleConfigurator.cc
@@ -116,7 +116,7 @@ simtime_t EagerGateScheduleConfigurator::computeStartOffsetForPathFragments(Inpu
             if (!strcmp(pathFragment->networkNodes.front()->module->getFullName(), std::get<0>(todo).c_str())) {
                 simtime_t nextGateOpenTime = std::get<1>(todo);
                 std::vector<std::string> extendedPath = std::get<2>(todo);
-                for (int i = 0; i < pathFragment->networkNodes.size() - 1; i++) {
+                for (size_t i = 0; i < pathFragment->networkNodes.size() - 1; i++) {
                     auto networkNodeName = pathFragment->networkNodes[i]->module->getFullName();
                     extendedPath.push_back(networkNodeName);
                     auto networkNode = getParentModule()->getSubmodule(networkNodeName);
@@ -133,7 +133,7 @@ simtime_t EagerGateScheduleConfigurator::computeStartOffsetForPathFragments(Inpu
                     simtime_t gateOpenDuration = transmissionDuration;
                     simtime_t gateOpenTime = nextGateOpenTime;
                     simtime_t gateCloseTime = gateOpenTime + gateOpenDuration;
-                    for (int i = 0; i < interfaceSchedule.size(); i++) {
+                    for (size_t i = 0; i < interfaceSchedule.size(); i++) {
                         if (interfaceSchedule[i].gateCloseTime + interFrameGap <= gateOpenTime || gateCloseTime + interFrameGap <= interfaceSchedule[i].gateOpenTime)
                             continue;
                         else {
@@ -187,7 +187,7 @@ void EagerGateScheduleConfigurator::addGateSchedulingForPathFragments(Input::Flo
             if (!strcmp(pathFragment->networkNodes.front()->module->getFullName(), std::get<0>(todo).c_str())) {
                 simtime_t nextGateOpenTime = std::get<1>(todo);
                 std::vector<std::string> extendedPath = std::get<2>(todo);
-                for (int i = 0; i < pathFragment->networkNodes.size() - 1; i++) {
+                for (size_t i = 0; i < pathFragment->networkNodes.size() - 1; i++) {
                     auto networkNodeName = pathFragment->networkNodes[i]->module->getFullName();
                     extendedPath.push_back(networkNodeName);
                     auto networkNode = getParentModule()->getSubmodule(networkNodeName);
@@ -204,7 +204,7 @@ void EagerGateScheduleConfigurator::addGateSchedulingForPathFragments(Input::Flo
                     simtime_t gateOpenDuration = transmissionDuration;
                     simtime_t gateOpenTime = nextGateOpenTime;
                     simtime_t gateCloseTime = gateOpenTime + gateOpenDuration;
-                    for (int i = 0; i < interfaceSchedule.size(); i++) {
+                    for (size_t i = 0; i < interfaceSchedule.size(); i++) {
                         if (interfaceSchedule[i].gateCloseTime + interFrameGap <= gateOpenTime || gateCloseTime + interFrameGap <= interfaceSchedule[i].gateOpenTime)
                             continue;
                         else {
